% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xval.R
\name{aaphcxval}
\alias{aaphcxval}
\title{Compute a retrospective hindcast cross-validation of AAP indices}
\usage{
aaphcxval(stock, indices, control, nyears = 5, nsq = 3, pin = NULL)
}
\arguments{
\item{stock}{Input FLStock object.}

\item{indices}{Input FLIndices object.}

\item{nyears}{Number if years for retrospective, defaults to 5.}

\item{nsq}{Number of years for average biology and selectivity, defaults to 3.}

\item{pin}{dsata.frame of parameter estimates to use as starting values.}
}
\value{
A list containing elements 'stocks', of class \emph{FLStocks}, and
'indices', a list of \emph{FLIndices} objects. See details for the structure of this list.
}
\description{
The output of \code{aaphcxval} consist of a list with two elements, named 'stocks'
and 'indices'. The first is an object of class \code{FLStocks}, each a peel from
the restrospective run. The second element is a list of \code{FLIndices} object.
The first \code{FLIndices} object, named 'data', is a copy of the input 'indices'
argument, with the additioned \code{catch.n} slot, if originally missing. The next
element, named as the final year of the data set, contains the naive prediction
of the input \code{FLIndices}, while the remaining elements are the result of a
hindcast prediction of the relevant indices, those within the year range of
as set ny \code{nyears}.
}
\examples{
data(sol4)
cd.command()
sxval <- aaphcxval(sol4 + fit, indices, control=control(fit),
  pin=stdfile(fit))
plotXval(sxval$indices)
plot(window(sxval$stocks, start=2005),
  metrics=list(SSB=ssb, F=fbar, Recruits=rec))
}
