% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{FLCLASS}
\alias{FLCLASS}
\alias{FLCLASS-methods}
\alias{FLCLASS-class}
\title{A class for}
\description{
Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque eleifend
odio ac rutrum luctus. Aenean placerat porttitor commodo. Pellentesque eget porta
libero. Pellentesque molestie mi sed orci feugiat, non mollis enim tristique.
Suspendisse eu sapien vitae arcu lobortis ultrices vitae ac velit. Curabitur id
}
\section{Slots}{

\describe{
\item{SLOT}{Neque porro quisquam est qui dolorem ipsum (\code{SLOTCLASS}).}
}
}

\section{Validity}{


\describe{
\item{VALIDITY}{Neque porro quisquam est qui dolorem ipsum.}
}

You can inspect the class validity function by using
\code{getValidity(getClassDef('FLCatch'))}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing but not stored in any slot.
}

\section{Methods}{

Methods exist for various calculations based on values stored in the class:

\describe{
\item{METHOD}{Neque porro quisquam est qui dolorem ipsum.}
}
}

\examples{

}
\seealso{
\link{FLComp}
}
\author{
The FLR Team
}
\keyword{classes}
