% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{\%pkgname\%}
\alias{\%pkgname\%}
\alias{\%pkgname\%-package}
\title{What The Package Does}
\description{
Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque eleifend
odio ac rutrum luctus. Aenean placerat porttitor commodo. Pellentesque eget porta
libero. Pellentesque molestie mi sed orci feugiat, non mollis enim tristique.
Suspendisse eu sapien vitae arcu lobortis ultrices vitae ac velit. Curabitur id
}
\examples{
\dontrun{FLPKG()}
}
\author{
%USER% \email{%EMAIL%, duplicate ampersand}

Maintainer: %USER% \email{%EMAIL%, duplicate ampersand}
}
